make every directory and ts/tsx file `kebab-case`
