+++ ui/app/api/chat-with-openai-streaming/route.ts
