+++ ui/app/api/get-image-description-openai/route.ts
