+++ ui/app/api/stable-video-diffusion/route.ts
