+++ ui/app/api/suggest-frontend-component/route.ts
