+++ ui/app/components/input-with-button.tsx
@@ -1,7 +1,7 @@
 'use client';
 import tw from 'tailwind-styled-components';
 import { IoSend } from 'react-icons/io5';
-import { stackDB } from '../stacks/stackDB';
+import { stackDB } from '../stacks/stack-db';
 import { useRouter } from 'next/navigation';
 
 import { useFormStatus } from 'react-dom';
