+++ ui/app/components/stacks/create-ai-canvas.tsx
