+++ ui/public/stack-pictures/chat-with-openai-streaming-helicone.png
