+++ ui/public/stack-pictures/create-ai-canvas.png
