+++ ui/public/stack-pictures/stable-video-diffusion.png
