+++ ui/public/stack-pictures/use-openai-assistant.png
