+++ ui/public/stacks/chat-with-openai-streaming/route.ts
