+++ ui/public/stacks/create-ai-canvas.tsx
