+++ ui/public/stacks/get-image-description-openai/route.ts
