+++ ui/public/stacks/stable-video-diffusion/route.ts
