+++ ui/scripts/sync-stacks.ts
